CH01-ex07 Reversed 90% of the PE function
In order to get started on this one I had to look at the solution so that
I could make sense of what was going on. What I could have done instead,
was looking up immediate values such as '0x3C' that can either be sizeof()
structures, offsets or options. A more aggressive use of the Internet
should help me when stuck at the beginning of challenges (most of the
time). I still have to figure out what does the function return and what
is it's type. It seems that <windows.h> is often used for low-level
interaction in general (i.e. malware and such)