Add sample code from Linux Device Drivers 3rd Ed.
The previous commit modified the README file stating the entire
repository would be reorganized. This is the first commit containing
code that is not from Bruce Dang's book.

Aside from the code provided with the book, there are folders containing
code manually wrote by me as I go through the book.
Changes to be committed:
    new file:   linux-device-drivers/chapter_01/Makefile
    new file:   linux-device-drivers/chapter_01/hello-world.c
    new file:   linux-device-drivers/ldd2-samples-1.0.1/Changes
    new file:   linux-device-drivers/ldd2-samples-1.0.1/LICENSE
    new file:   linux-device-drivers/ldd2-samples-1.0.1/Makefile
    new file:   linux-device-drivers/ldd2-samples-1.0.1/Makefile.alpha
    new file:   linux-device-drivers/ldd2-samples-1.0.1/Makefile.arm
    ...